Streamline/fix compilation and library support
I used some code from the Bare Arduino Project[1], discarding and
modifying where I saw fit. Libraries are stored in this git repo now,
and in fact part of the compilation issue was getting eHealthDisplay.h
to compile properly - a patch was applied to fix it, so storing the
libraries locally is best. To access the local libraries, a patched
version of Arduino Makefile[2] is required, so I use that in a
submodule instead of the one provided by apt.

Compilation is working - it would be nice to get some tests and the core
functionality in there too

[1]: https://github.com/ladislas/Bare-Arduino-Project
[2]: https://github.com/WeAreLeka/Arduino-Makefile